code/requirements.txt
h5py
imageio
Pillow
torch
torchvision
matplotlib
scikit-learn
tqdm